+++ rotkehlchen/chain/evm/decoding/efp/constants.py
@@ -1,0 +1,3 @@
+from typing import Final
+
+CPT_EFP: Final = 'efp'