+++ rotkehlchen/chain/optimism/modules/efp/decoder.py
@@ -1,0 +1,25 @@
+from typing import TYPE_CHECKING
+
+from rotkehlchen.chain.evm.decoding.efp.decoder import EfpCommonDecoder
+from rotkehlchen.chain.evm.types import string_to_evm_address
+
+if TYPE_CHECKING:
+    from rotkehlchen.chain.evm.decoding.base import BaseDecoderTools
+    from rotkehlchen.chain.optimism.node_inquirer import OptimismInquirer
+    from rotkehlchen.user_messages import MessagesAggregator
+
+
+class EfpDecoder(EfpCommonDecoder):
+
+    def __init__(
+            self,
+            optimism_inquirer: 'OptimismInquirer',
+            base_tools: 'BaseDecoderTools',
+            msg_aggregator: 'MessagesAggregator',
+    ) -> None:
+        super().__init__(
+            evm_inquirer=optimism_inquirer,
+            base_tools=base_tools,
+            msg_aggregator=msg_aggregator,
+            list_records_contract=string_to_evm_address('0x4Ca00413d850DcFa3516E14d21DAE2772F2aCb85'),
+        )